+++ demos/gemini-xrobject/main.js
@@ -35,7 +35,7 @@ options.sound.speechRecognizer.playSimulatorActivationSounds = true;
 options.ai.enabled = true;
 options.ai.gemini.enabled = true;
 options.ai.gemini.model = 'gemini-2.5-flash';
-options.world.objects.backendConfig.activeBackend = 'mediapipe';
+options.world.objects.backendConfig.activeBackend = 'gemini';
 options.world.objects.showDebugVisualizations = false;
 options.setAppTitle('Gemini XR-Objects');
 options.setAppDescription(
